+++ lib/wechat_kit.dart
@@ -5,3 +5,4 @@ export 'src/model/req.dart';
 export 'src/model/resp.dart';
 export 'src/wechat.dart';
 export 'src/wechat_constant.dart';
+export 'src/wechat_kit_platform_interface.dart';